Authorize

+++ SocialMediaApp/Controllers/GroupsController.cs
@@ -1,7 +1,9 @@
 ﻿using Humanizer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Hosting;
 using SocialMediaApp.Data;
 using SocialMediaApp.Models;
 using System.Net.NetworkInformation;
@@ -26,6 +28,7 @@ namespace SocialMediaApp.Controllers
             _userManager = userManager;
             _roleManager = roleManager;
         }
+        [Authorize(Roles = "User,Moderator,Admin")]
         public IActionResult Index()
         {
             var grupuri = from item in db.Groups
@@ -54,8 +57,9 @@ namespace SocialMediaApp.Controllers
                         return View(group);
                     }
                 }*/
+        [Authorize(Roles = "Moderator,Admin")]
         [HttpPost]
-        public async Task<IActionResult> New(Group group, IFormFile Image)
+        public async Task<IActionResult> New(Group group, IFormFile? Image)
         {
             if (User.Identity == null || !User.Identity.IsAuthenticated)
             {
@@ -63,13 +67,13 @@ namespace SocialMediaApp.Controllers
                 return View(group);
             }
 
-/*            group.UserId = _userManager.GetUserId(User);
+            /*            group.UserId = _userManager.GetUserId(User);
 
-            if (string.IsNullOrEmpty(group.UserId))
-            {
-                ModelState.AddModelError("UserId", "Unable to determine the user ID.");
-                return View(group);
-            }*/
+                        if (string.IsNullOrEmpty(group.UserId))
+                        {
+                            ModelState.AddModelError("UserId", "Unable to determine the user ID.");
+                            return View(group);
+                        }*/
 
             if (Image != null && Image.Length > 0)
             {
@@ -103,6 +107,7 @@ namespace SocialMediaApp.Controllers
             }
             return View(group);
         }
+        [Authorize(Roles = "User,Moderator,Admin")]
         public IActionResult Show(int id)
         {
             Group group = db.Groups.Include("Posts")
@@ -126,6 +131,11 @@ namespace SocialMediaApp.Controllers
         public IActionResult Delete(int id)
         {
             var group = db.Groups.Find(id);
+            if (!(User.IsInRole("Moderator") || User.IsInRole("Admin")))
+            {
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra unui articol care nu va apartine";
+                return RedirectToAction("Index");
+            }
             db.Groups.Remove(group);
             db.SaveChanges();
             return RedirectToAction("Index");
@@ -133,7 +143,15 @@ namespace SocialMediaApp.Controllers
         public ActionResult Edit(int id)
         {
             Group grup = db.Groups.Find(id);
-            return View(grup);
+            if (User.IsInRole("Moderator") || User.IsInRole("Admin"))
+            {
+                return View(grup);
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra unui articol care nu va apartine";
+                return RedirectToAction("Index");
+            }
         }
 
         /* [HttpPost]
@@ -157,10 +175,15 @@ namespace SocialMediaApp.Controllers
              }
          }*/
         [HttpPost]
-        public async Task<IActionResult> Edit(int id, Group requestGroup, IFormFile Image)
+        public async Task<IActionResult> Edit(int id, Group requestGroup, IFormFile? Image)
         {
             Group group = db.Groups.Find(id);
-            try
+            if (!(User.IsInRole("Moderator") || User.IsInRole("Admin")))
+            {
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra unui articol care nu va apartine";
+                return RedirectToAction("Index");
+            }
+            if(ModelState.IsValid)
             {
                 group.Nume = requestGroup.Nume;
                 group.Descriere = requestGroup.Descriere;
@@ -191,7 +214,7 @@ namespace SocialMediaApp.Controllers
                 TempData["message"] = "Articolul a fost modificat";
                 return RedirectToAction("Index");
             }
-            catch (Exception)
+            else
             {
                 return View(requestGroup);
             }
